BUG: vtkMRMLUnitNodeTest1 use size_t instead of int
This fixes the warnings for linux brought by r22053

  /Slicer-TR/Libs/MRML/Core/Testing/vtkMRMLUnitNodeTest1.cxx: In function ‘bool TestScenesUnitNodeID(vtkMRMLScene*)’:
  /Slicer-TR/Libs/MRML/Core/Testing/vtkMRMLUnitNodeTest1.cxx:86: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
  /Slicer-TR/Libs/MRML/Core/Testing/vtkMRMLUnitNodeTest1.cxx:92: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
  /Slicer-TR/Libs/MRML/Core/Testing/vtkMRMLUnitNodeTest1.cxx: In function ‘bool TestUnitNodeAttribute(vtkMRMLScene*)’:
  /Slicer-TR/Libs/MRML/Core/Testing/vtkMRMLUnitNodeTest1.cxx:121: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]

And the same kind of warning in the logic test.

See Issue #1694

svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=22061
git-svn-id: http://svn.slicer.org/Slicer4/trunk@22061 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRML/Core/Testing/vtkMRMLUnitNodeTest1.cxx
@@ -26,7 +26,10 @@
 // VTK includes
 #include <vtkSmartPointer.h>
 
-const int NUMBER_OF_UNITS = 5;
+// STD includes
+#include <vector>
+
+const size_t NUMBER_OF_UNITS = 5;
 const char* UNITS[NUMBER_OF_UNITS][2] = {{"length", "m"},
                                          {"length", "km"},
                                          {"energy", "J"},
@@ -63,7 +66,7 @@ int vtkMRMLUnitNodeTest1(int , char * [] )
 vtkMRMLScene* CreatePopulatedScene()
 {
   vtkMRMLScene* scene = vtkMRMLScene::New();
-  for (int i = 0; i < NUMBER_OF_UNITS; ++i)
+  for (size_t i = 0; i < NUMBER_OF_UNITS; ++i)
     {
     vtkSmartPointer<vtkMRMLUnitNode> unit
       = vtkSmartPointer<vtkMRMLUnitNode>::New();
@@ -89,7 +92,7 @@ bool TestScenesUnitNodeID(vtkMRMLScene* scene)
     return false;
     }
 
-  for (int i = 0; i < units.size(); ++i)
+  for (size_t i = 0; i < units.size(); ++i)
     {
     std::string id = "vtkMRMLUnitNode";
     id += UNITS[i][1];
@@ -118,7 +121,7 @@ bool TestUnitNodeAttribute(vtkMRMLScene* scene)
 
   bool foundQuantityAttribute = false;
   std::vector<std::string> attributes = unit->GetAttributeNames();
-  for (int i = 0; i < attributes.size(); ++i)
+  for (size_t i = 0; i < attributes.size(); ++i)
     {
     if (attributes[i].compare("Quantity") == 0)
       {
